+++ Userland/Libraries/LibJS/Runtime/Set.h
@@ -7,30 +7,11 @@
 #pragma once
 
 #include <AK/HashTable.h>
-#include <LibJS/Runtime/BigInt.h>
 #include <LibJS/Runtime/GlobalObject.h>
 #include <LibJS/Runtime/Object.h>
 #include <LibJS/Runtime/Value.h>
 
 namespace JS {
-
-struct ValueTraits : public Traits<Value> {
-    static unsigned hash(Value value)
-    {
-        VERIFY(!value.is_empty());
-        if (value.is_string())
-            return value.as_string().string().hash();
-
-        if (value.is_bigint())
-            return value.as_bigint().big_integer().hash();
-
-        return u64_hash(value.encoded()); // FIXME: Is this the best way to hash pointers, doubles & ints?
-    }
-    static bool equals(const Value a, const Value b)
-    {
-        return same_value_zero(a, b);
-    }
-};
 
 class Set : public Object {
     JS_OBJECT(Set, Object);
